% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComPlot_func_v1.R
\name{comPlot}
\alias{comPlot}
\title{Plot output from multiple saved Habplan flow files}
\usage{
comPlot(flow.data.1, flow.data.2, flow.data.3, nyear)
}
\arguments{
\item{flow.data.1}{Flow file saved by Habplan}

\item{flow.data.2}{Flow file saved by Habplan}

\item{flow.data.3}{Flow file saved by Habplan}

\item{nyear}{Number of years or time periods interested in (have data for)}
}
\value{
A figure comparing multiple flows across time period
}
\description{
Creates a figure which compares the flow outputs across time periods for multiple flow files
}
\examples{
#Read in the example flows:
flow1 <- read.csv("./saveFlow1", sep="")
flow2 <- read.csv("./saveFlow2", sep="")
flow3 <- read.csv("./saveFlow3", sep="")

#Run function with new flows
comPlot(flow.data.1 = flow1, flow.data.2 = flow2,
        flow.data.3 = flow3, nyear = 35)

}
