% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowPlot_func_v1.R
\name{flowPlot}
\alias{flowPlot}
\title{Plot output from a single saved Habplan flow file}
\usage{
flowPlot(flow.data, nyear)
}
\arguments{
\item{flow.data}{Flow file saved by Habplan}

\item{nyear}{Number of years or time periods interested in (have data for)}
}
\value{
A figure comparing multiple flows across time period
}
\description{
Creates a figure to visualize flow outputs across time periods for a single flow file
}
\examples{
#Read in the example flows:
flow1 <- read.csv("./saveFlow1", sep="")

#Input the flow file into the function, and number of years
flowPlot(flow.data = flow1, nyear = 35)

}
