% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HabBio2_func_v1.R
\name{HabBio2}
\alias{HabBio2}
\title{Convert data frames to Habplan-suitable flow files}
\usage{
HabBio2(std.data, std.info)
}
\arguments{
\item{std.data}{Forest stand, regime, and outcome information}

\item{std.info}{Additional stand information}

\item{col}{The column number for the data interested in from std.data}

\item{nyear}{Number of years or time periods interested in (have data for)}

\item{HSI}{Threshold value of HSI for species of interest (0-0.99)}
}
\value{
Saves a .dat file in the working directory ready for import to Habplan
}
\description{
Creates flow files from csv input data
}
\examples{
#Read in stand and regime information
std.data <- read_csv("./fvs_results.csv")

#Read in additional stand information - acreage, id, and description
std.info <- read_csv("./Stand_info.csv")

#Look at the column headings to decide which column has flow results
colnames(new.data)

#From the column names, we are interested in "HSI", which is
#column number 37. We can input this into our function.

#Run function
rcw.flow <- habConvert(std.data = new.data, std.info = std.info, col = 37,
                       nyear = 35, HSI = 0.7)

}
