% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HabBlock_func_v2.R
\name{HabBlock}
\alias{HabBlock}
\title{Create block size constraint data}
\usage{
HabBlock(std.data, std.info, site.shp, block.title)
}
\arguments{
\item{std.data}{Forest stand, regime, and outcome information}

\item{std.info}{Additional stand information}

\item{site.shp}{SpatVector file of study area (forest stands)}

\item{block.title}{Name of the to-be saved .txt file}
}
\value{
Saves a .dat file in the working directory ready for import to Habplan
}
\description{
Creates flow files from csv input data
}
\examples{
#Read in stand and regime information
std.data <- read_csv("./fvs_results.csv")

#Read in additional stand information - acreage, id, and description
std.info <- read_csv("./Stand_info.csv")

#Read in stand shapefiles
site.shp <- readOGR(dsn = "./shapefiles/Stands2_Shapefile/Stands_final.shp")

#Convert to a SpatVector
site.shp <- vect(site.shp)

#Run function
HabBlock(std.data = std.data, std.info = std.info, site.shp = site.shp,
block.title = "block2")

}
