% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HabSpace_func_v1.R
\name{HabSpace}
\alias{HabSpace}
\title{Landscape metric analyses on study site using Habplan output file}
\usage{
HabSpace(
  site.shp = site.shp,
  flow = flow,
  nyear = nyear,
  mode = "terrestrial",
  dist = 500,
  level = "patch"
)
}
\arguments{
\item{site.shp}{SpatVector file of study area (forest stands)}

\item{flow}{Output flow file of interest (typically habitat area)}

\item{nyear}{Number of years or time periods interested in (have data for)}

\item{mode}{"terrestrial" or "avian" depending on movement type of study species}

\item{dist}{Dispersal distance of species if no habitat exists. I.e., the possible barrier width that can be traversed by an avian species.}

\item{level}{Levels based on the "Landscapemetrics" package - "patch", "landscape", or "class"}
}
\value{
FIgures of landscapes over time, csv files of landscape metrics, shapefiles of all stands over time
}
\description{
Applies various landscape metric analyses on study site using Habplan output file
}
\examples{
#Read in stand shapefiles
site.shp <- readOGR(dsn = "./shapefiles/Stands2_Shapefile/Stands_final.shp")

#Read in flow file from Habplan run
flow <- read.csv("./saveFlow1.dat")

##Important: The shapefile needs a data column called "StdID"##

#We will add the shapefile of stands, and the flow of choice to the
#function. As before, we will set the nyear. We now have two more options
#with this function, mode and dist.

#There are two modes to choose from, terrestrial or avian - depending
#on the species the HSI has been built around. If avian is chosen, the
#dist setting sets the distance in meters (or the unit of the shapefile)
#that the species of interest can comfortably travel to reach neighboring
#habitat patches. If the user has a terrestrial species which may still
#bridge gaps if the distance is short enough, e.g. crossing a road, then
#the avian option should be selected, and a smaller dist assigned.

#Want to create a shapefile output for each year, to look at change
#of regime for each stand over time.

test.x <- HabSpace(site.shp = site.shp, flow = flow, nyear = 35,
                 mode = "terrestrial", dist = 500, level = "landscape")


}
