% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSIcalc_func_v1.R
\name{HSIcalc}
\alias{HSIcalc}
\title{Habitat Suitability Index calculation}
\usage{
HSIcalc(std.data, equation, logistic = T)
}
\arguments{
\item{std.data}{Forest stand, regime, and outcome information}

\item{equation}{User-defined equation for calculating HSI}

\item{logistic}{TRUE/FALSE - if True, runs a logistic function on values (assigns HSI values between 0-1)}
}
\value{
The input stand data with an additional HSI column of values
}
\description{
Calculates the Habitat Suitability Index (HSI) for a species of interest
}
\examples{
#Read in stand and regime information
std.data <- read_csv("./fvs_results.csv")

#Create function to apply equation for HSI
Define the slope and y-intercept of the linear equation
hsi.func <- function(x = std.data$BA, m = -0.02, b = 2) {
return(m * x + b)
}

The function will create a new column with the HSI values
new.data <- HSIcalc(std.data, hsi.func)

#Summarise the HSI data
summary(new.data$HSI)
#Plot data against BA to show relationship
ggplot(data = new.data) +
geom_point(aes(x = BA, y = HSI, color = HSI),
          size = 3) +
          scale_color_viridis_c(option = "plasma") +
          theme_classic()

}
