% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandSched_func_v1.R
\name{standSched}
\alias{standSched}
\title{Save top Habplan schedules to shapefile}
\usage{
standSched(site.shp)
}
\arguments{
\item{site.shp}{SpatVector file of study area (forest stands)}
}
\value{
A .shp file with each stand regime included
}
\description{
Saves the shedules provided by Habplan to the study site shapefiles
}
\examples{
#Read in stand shapefiles
site.shp <- readOGR(dsn = "./shapefiles/Stands2_Shapefile/Stands_final.shp")

#Run function
new.shp <- standSched(site.shp = site.shp)

}
